Add coverage via nyc.
Include appropriate annotations to avoid inlined functions within
convertMarkdownToMocha used to generate code from instrumentation.

Note that one test, specifically the application of cleanStackTrace,
inlines a chunk of code from an external file. When run by coverage
the inlined version has instrumentation in it thus causing the test
to fail. Most of the solutions to this a somewhat nasty - opt to
sidestep the problem for now.