Implement a CLI and config file support.
This commit adds the infrastructure required to load options from
a file. Expose a bin file that consumes this to enable generation
to occur for multiple configurations.

The configurations files themselves as JS files (primarily to allow
specifying function to change e.g. a template function to influence
how the output HTML is treated.