Implement capturing console output.
Add and outputCapture key that selects the style of output capturing
that will be performed. Support a capture type of "console" that
causes the console passed to evaluated snippets to be swapped out
for an InspectedConsole.

Instances of InspectedConsole are API compatible with the standard
console. Internally, output is buffered and any non-string output
is inspect()ed by Unexpected. This buffered output serialised back
out as snipped output and written to the output file.