Implement a mechanism to rethrow certain error during revaluation.
Due to the feature that errors thrown by the snippets are captured,
it cam be difficult to hunt down the cause of larger issues. Choose
to treat RefereneceError exceptions are reasons to bail out of the
evaluation - meaning that rather than it being serialised it will
cause a generation time event which is much easier to deal with.