Split out single snippet eval & use structured output representation.
Factoring this logic out is a big aid to readability and is a good
preparatory step to being able to take more actions per-snippet.

Use the separation to return an output structure from evaluation
which is then attached as a whole to each snippet. Within callers
this means multiple property checks can collapse to linear paths.