Exhaust all the snippets in a file before reporting evaluation error.
Split the evaluation errors into a snippet and file error. Store any
error generated for a snippet in a map but continue processing other
snippets. After evaluating all the snippets for a file, consult the
map and if any errors were recorded create a combined file error.

This allows us to output all the failures for a file in one go which
enables the user to easily address all the issues that were found.
The changes also pave the way for best effort file evaluation.