Add processing of a single file in-place through into the CLI.
Implement an updateFile() counterpart to processFile() which deals
only with updating the source file and has no knowledge of a target
file or directory.

Use this within a new file command method in the CLI to support
processing a single file - either generating the output or writing
it in-place via updateFile().

Modify the bin file to be aware of a positional file argument. If
supplied it will invoke the single file CLI function and otherwise
it will proceed with directory processing version as before.