Rearrange error handling to allow re-use of errorToInfo().
In the case of single file processing, it would be somewhat ideal to
get the same detailed breakdown of which snippets failed and why if
an error were to occur.

Use the code previously written to serialise a FileEvaluationError
for debugging output and promote it to shared logic. Hook is into
the bin file to have an error being throw be all that is necessary
to have it pretty printed to the console.