Begin pushing capture mode down into evaluation.
Use the snippet flags mechanism to reconfigure the way output it
captured. For each snippet, use options supplied by the caller to
attach flags that cause the different modes to take effect. Ensure
this is done in such a way that the evaluated flags are not leaked
to the outside.