Allow specifying the capture type on a per-snippet basis.
Move grabbing of the console to the point where a single snippet is
about to be run in evaluateSnippets(). Do this on the basis of which
flags are set. After the previous commit a localised flag object was
already created which means the environment we prepare is affected
by any flags on the snippet.