Allow passing a directory to operate on via the command line.
Add a new CLI function that does a stat on its argument and executes
the correct method based on whether it sees a directory or a file.

Split the inplace option from update and consistently define it as
meaning to avoid the target path completely and instead make changes
to the source files only. In this way, when combined with being able
to pass a directory, we can use the tool to update all the examples
within a documentation directory with no configuration.

While here, ensure that a file source path is validated.