Fix missing quotes around strings logged to the console.
Special handling surrounded strings logged to the console to ensure
that output intended for stderr would coloured by the error style.
Doing this however bypassed the style that would quote the string
and highlight any special characters that it contained.

Copy the style definition inline and modify to allow selection of
the style to use conditional upon the output file descriptor.