Implement TypeScript support.
Make "typescript" a second supported evaluation language. Alter the
snippet extraction code to recognise "typescript" language and allow
"ts" as an alias. Rearrange the evaluation code to, upon discovery
of a TypeScript snippet, to set an transpilation function shipped
as part of the tool that invoked the TS compiler on snippets.

In order to make use of TypeScript, require a base level of explicit
choice in the form of requiring a tsconfig to be specified.