Wrap snippets only once at the point of transpilation.
Snippets that rely on being able to use the return statement need to
have the wrapping function applied prior to transpilation otherwise
their compilation will fail with bad syntax.

Prior to this commit this was the intention, but it was incorrectly
applied twice both when the snippets were transpiled but also during
evaluation. Wrapping at transpile time is the right place so move it
there and ensure it is only done once. Also fix the case of global
capture settings and pre-snippet capture override - these were not
determining the flags correctly as is done by the evaluation code,
so account for this by calling the same preparation code and using
the resulting flags as the basis of the wrapping.