Implement a require option for a preamble to inject per file.
Allow a require option to be supplied on the command line - the
file pointed will have its contents pulled in and executed as a
preamble to every evaluated markdown file.

Note that the preamble we load itself may contain require() calls
and these must be resolved relative to the file whose content we
load if we are to handle local imports within projects. In order
to achieve this we record both the content of the file but in
addition it's path and create a new require function for that
path at the point of preamle execution.

While here, address allowing snippets to include their own local
utility modules. For the remaining snippets in file, use a require
that is configured with the directory containing the snippet file
as the working directory. Exercise modules local to the snippets.