Repurpose the async flag to allow use of the await keyword.
Given the pervasive use of promises and the default capture mode
being the return value, simply check for a promise being returned
from snippets and wait for its completion if supplied.

This behaviour frees up the async flag which is redefined as a
hint to cause wrapping snippets by an async function thus allowing
use of the await keyword within the code block.