Fix an issue around the serialisation of comments.
The base index of the snippet block within the file having the
comment length added on which caused comments to be present in
the html/inlined output formats. In addition, when it came to
markdown output, these were already being taken into account
when writing the output string so they would end up duplicated.

Correct the index thus removing comments except for markdown.