Make freshContext handle fileGlobals correctly.
File globals were previously being created by Markdown and handed
down as an object to the evaluation logic. This meant that when a
fresh context was requested they were not being reinitialised.

Relocate and tie the creation of file globals into the code that
prepares the globals within evaluateSnippets. The same logic is
called at the point a flag requesting fresh context is processed,
meaning that they become recreated. Demonstrate this is the case
by adding an example that calls an Unexpected API (addType) that
throws in the case where we don't use a fresh clone. Verify this
example works both with a require()d expect and as a file global.