Remove nowrap and add flag for globals to persist into fresh context.
One of the principle ideas of this documentation was that, within a
file, we act as if the entire thing is a manuscript to be read from
top to bottom. Isolating each block and forcing each case desiring
persisting made what we'd consider standard behaviour very verbose.

Return to the original premise and allow definitions earlier in a
file to remain until the end. Note that fresh contexts present an
issue, wherein they should be a totally fresh global, so instead
add a flag to demarcate definitions that should survive into them.