Redo the fix committed in b128d5e.
Clearing the recorded expect on the Markdown instance means we
cannot later render coloured HTML output when we transition from
evaluation of the snippets to rendering them. But we still need
to make sure that we will have the ability to pretty print using
the customised expect - solve this by checking for the necessary
plugin and creating a suitable clone for that purpose if not.