Always replace explicitly listed globals during cleanup.
In a previous commit the restoration of internally replaced globals
had become conditional - the consequence of this was a instance of
the inspected console could be left attached to the global which
could mean that later logging attempts as part of exiting the tool
would not make it to the terminal. Address this by always restoring.