Use mocha reporters to output validation mode results to the console.
Remove the the Evaldown instance validate flag and instead completely
split the --validate option codepath from the processing code. The two
are now distinct methods called on the instance with the shared piece
being the logic to perform evaluation of the snippets within a file.

Leverage the split to rework the validation code as a moch-esque event
emitter by porting the mocha reporter glue from fugl - each validated
file is recorded as an explicit validation result which is emitted.

As part of this, the --reporter and --ci options are implemented but
resricted to use with --validate; HTML reports are written to disk.