Rework error architecture and surface require errors as per-snippet.
Instead of always throwing from prepareFile() opt to record any errors
that were encountered - either by check() or evaluate() - and allow a
caller to opt-out of the throw.

Use this functionality when validating snippets to return these "early"
errors and mix then into the overall results emitted for each snippet.

The practical effect of this is to enable surfacing a module require
error that occurs in a snippet unified within the per-snippet output.