Fix a global environment leak with skipped snippets.
With a capture mode of "console", each snippet whose environment is
prepared will be provided with the a console instrumented such that
logged messages can be inspected. This preparation was also done for
output snippets, but these are not executed and thus cleanup() isn't
called for them resulting in it being leaked. Fix this by reordering
the checks so preparation is not done for output snippets.