Rationalise handling of the expect for output inspection.
It turns out that ever since the introduction of logic to either
grab a global unexpected and clone/extend if it exists, all the
codepaths must call setExpect() method. This means we could avoid
the base expect thing if we just dealt with the absence of an expect
global in that method. Do this and remove the whole concept.

While here also note that the same mechanism could be reused to grab
a fresh expect when doing validation.. make that change too and thus
rip out the unused createExpect() function.

Note about createExpect(): the function contained some extra option
handling logic to, theoretically, allow custom configuration of the
expect instance. The most important of these was the format, but
that responsibility was subsumed by explicitly the output format
when the snippets are serialised. It turns out callers don't pass
the other options (instead they pass the global wholesale instead)
and at least the "theme" option needs handling on a per-file basis
by interrogating metadata on global creation. Opt to pull it out.