Restore passing format and test override after 89efc54.
In order for an HTML version of the error message to be sent to be
sent to the corresponding mocha reporter the Evaldown output format
name is consulted during validation error serialisation. Pass down
this formatName from the Evaldown instance in an explicit way. Also
restore allowing thew output format override needed for one specific
test but use a separate property to sidestep any possible confusion.